internal/app: remove commented out slackRead route/method
Quickly realized we probably don't want messages readable in slack
(similar to the CLI) because that defeats the purpose of no credentials
in slack. For now you will be able to create one time URLs in slack and
the user you are sending it too will simply click on the link.